# example/swag-gin-demo/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是一个展示如何使用 ApiKeyAuth 的示例项目。",
        "title": "Gin Swagger Example API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/ping": {
            "get": {
                "description": "返回 pong 消息，无需鉴权",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "心跳测试 (公开)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/secure/data": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "需要提供 Header 中的 Authorization 才能访问",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Secure"
                ],
                "summary": "获取私密数据 (受保护)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "AppKeyAuth": {
            "type": "apiKey",
            "name": "X-APP-KEY",
            "in": "header"
        },
        "AppSecretAuth": {
            "type": "apiKey",
            "name": "X-APP-SECRET",
            "in": "header"
        }
    }
}